com/demo 폴더구조 삭제 Readme 프로젝트 전체 구조, 개요 설명문으로 변경
