+++ .gitignore
@@ -25,6 +25,7 @@ bin/
 out/
 !**/src/main/**/out/
 !**/src/test/**/out/
+/src/main/resources/application.properties
 
 ### NetBeans ###
 /nbproject/private/
@@ -35,3 +36,4 @@ out/
 
 ### VS Code ###
 .vscode/
+
