+++ src/main/java/SilkLoad/SessionConst.java
@@ -1,0 +1,5 @@
+package SilkLoad;
+
+public interface SessionConst {
+    public static final String LOGIN_MEBMER = "loginMember";
+}
