+++ src/main/java/SilkLoad/dto/LoginFormDto.java
@@ -1,0 +1,16 @@
+package SilkLoad.dto;
+
+import lombok.Data;
+
+import javax.validation.constraints.NotBlank;
+
+@Data
+public class LoginFormDto {
+
+    @NotBlank
+    private String loginId;
+
+    @NotBlank
+    private String password;
+
+}
