+++ src/main/java/SilkLoad/dto/MemberFormDto.java
@@ -1,0 +1,25 @@
+package SilkLoad.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+
+import javax.validation.constraints.NotBlank;
+
+
+@Getter
+@Setter
+@ToString
+public class MemberFormDto {
+
+    @NotBlank
+    private String loginId;
+
+    @NotBlank
+    private String name;
+
+    @NotBlank
+    private String password;
+
+
+}
