+++ src/main/java/SilkLoad/entity/Members.java
@@ -1,0 +1,40 @@
+package SilkLoad.entity;
+
+import lombok.*;
+
+import javax.persistence.*;
+import java.time.LocalDate;
+
+
+@Entity
+@Getter
+@Setter
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class Members {
+
+
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "MEMBER_ID")
+    private Long id;
+
+    @Column(nullable = false)
+    private String loginId;
+
+    @Column(nullable = false)
+    private String name;
+
+    @Column(nullable = false)
+    private String password;
+
+    private Integer ranks;
+
+    private Integer numberPurchase;
+
+    private LocalDate createDate = LocalDate.now();
+
+
+
+}
