+++ src/main/java/SilkLoad/interceptor/LoginCheckInterceptor.java
@@ -1,0 +1,55 @@
+package SilkLoad.interceptor;
+
+import SilkLoad.SessionConst;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.servlet.HandlerInterceptor;
+import org.springframework.web.servlet.ModelAndView;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+
+@Slf4j
+public class LoginCheckInterceptor implements HandlerInterceptor {
+
+    /**
+     * 인터셉터 처리 첫 부분
+     * 컨트롤러에 요청이 가기전에 실행되는 메소드
+     */
+    @Override
+    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
+
+        String requestURI = request.getRequestURI();
+
+        log.info("인증 체크 인터셉터 실행 {}", requestURI);
+        HttpSession session = request.getSession(false);
+
+        if (session == null || session.getAttribute(SessionConst.LOGIN_MEBMER) == null ) {
+            log.info("미인증 사용자 요청");
+
+            response.sendRedirect("/login?redirectURL=" + requestURI);
+            //컨트롤러에 요청 x
+            return false;
+        }
+        //컨트롤러 요청 O
+        return true;
+
+    }
+
+    /**
+     * preHandle을 통과하고 controller를 거친 후에 실행되는 메소드
+     */
+    @Override
+    public void postHandle(HttpServletRequest request, HttpServletResponse response, Object handler, ModelAndView modelAndView) throws Exception {
+        HandlerInterceptor.super.postHandle(request, response, handler, modelAndView);
+    }
+
+    /**
+     * preHandle에서 오류가 나던 안 나던 무조건 실행되는 부분
+     * preHandle 실행 후에 실행되는 부분
+     */
+    @Override
+    public void afterCompletion(HttpServletRequest request, HttpServletResponse response, Object handler, Exception ex) throws Exception {
+        HandlerInterceptor.super.afterCompletion(request, response, handler, ex);
+    }
+}
