+++ src/main/java/SilkLoad/repository/ProductRepository.java
@@ -1,0 +1,11 @@
+package SilkLoad.repository;
+
+import SilkLoad.entity.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProductRepository extends JpaRepository<Product, Long> {
+
+    
+}
