+++ src/main/java/SilkLoad/service/ProductService.java
@@ -1,0 +1,55 @@
+package SilkLoad.service;
+
+import SilkLoad.dto.ProductFormDto;
+import SilkLoad.entity.Category;
+import SilkLoad.entity.Members;
+import SilkLoad.entity.Product;
+import SilkLoad.repository.ProductRepository;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+@Slf4j
+public class ProductService {
+
+    private final ProductRepository productRepository;
+
+    public void save(ProductFormDto productFormDto, Members loginMember) {
+
+        String categoryName = productFormDto.getCategory();
+
+        Category category = categoryClassification(categoryName);
+
+
+        Product product = Product.builder()
+                .name(productFormDto.getName())
+                .instantPrice(productFormDto.getInstancePrice())
+                .auctionPrice(productFormDto.getAuctionPrice())
+                .imageNameSrc(productFormDto.getImageNameSrc())
+                .Explanation(productFormDto.getExplanation())
+                .createdDate(productFormDto.getCreatedDate())
+                .members(loginMember)
+                .build();
+
+        product.changeCategory(category);
+
+        productRepository.save(product);
+
+    }
+
+    private Category categoryClassification(String categoryName ) {
+
+        String[] splitCategory = categoryName.split(",");
+        Category category = Category.builder()
+                .first(splitCategory[0])
+                .second(splitCategory[1])
+                .build();
+
+        return category;
+
+    }
+
+
+}
