--- src/main/java/com/example/demo/SessionConst.java
@@ -1,5 +1,0 @@
-package com.example.demo;
-
-public interface SessionConst {
-    public static final String LOGIN_MEBMER = "loginMember";
-}
