--- src/main/java/com/example/demo/config/WebConfig.java
@@ -1,23 +1,0 @@
-package com.example.demo.config;
-
-import com.example.demo.interceptor.LoginCheckInterceptor;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-
-@Configuration
-public class WebConfig implements WebMvcConfigurer {
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(new LoginCheckInterceptor())
-                .order(1)
-                .addPathPatterns("/**")
-                .excludePathPatterns(
-                        "/", "/members/add", "/login", "/logout","/home",
-                        "/css/**", "/*.ico", "/error","/vendor/**", "/img/**","/js/**","/fonts/**"
-                );
-    }
-
-}
