--- src/main/java/com/example/demo/dto/LoginFormDto.java
@@ -1,16 +1,0 @@
-package com.example.demo.dto;
-
-import lombok.Data;
-
-import javax.validation.constraints.NotBlank;
-
-@Data
-public class LoginFormDto {
-
-    @NotBlank
-    private String loginId;
-
-    @NotBlank
-    private String password;
-
-}
