--- src/main/java/com/example/demo/dto/MemberFormDto.java
@@ -1,25 +1,0 @@
-package com.example.demo.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-import javax.validation.constraints.NotBlank;
-
-
-@Getter
-@Setter
-@ToString
-public class MemberFormDto {
-
-    @NotBlank
-    private String loginId;
-
-    @NotBlank
-    private String name;
-
-    @NotBlank
-    private String password;
-
-
-}
