--- src/main/java/com/example/demo/entity/Members.java
@@ -1,40 +1,0 @@
-package com.example.demo.entity;
-
-import lombok.*;
-
-import javax.persistence.*;
-import java.time.LocalDate;
-
-
-@Entity
-@Getter
-@Setter
-@ToString
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Members {
-
-
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "MEMBER_ID")
-    private Long id;
-
-    @Column(nullable = false)
-    private String loginId;
-
-    @Column(nullable = false)
-    private String name;
-
-    @Column(nullable = false)
-    private String password;
-
-    private Integer ranks;
-
-    private Integer numberPurchase;
-
-    private LocalDate createDate = LocalDate.now();
-
-
-
-}
