--- src/main/java/com/example/demo/repository/MemberRepository.java
@@ -1,14 +1,0 @@
-package com.example.demo.repository;
-
-import com.example.demo.entity.Members;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface MemberRepository extends JpaRepository<Members, Long>{
-
-    Optional<Members> findByLoginId(String LoginId);
-
-}
