--- src/main/java/com/example/demo/repository/ProductRepository.java
@@ -1,11 +1,0 @@
-package com.example.demo.repository;
-
-import com.example.demo.entity.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ProductRepository extends JpaRepository<Product, Long> {
-
-    
-}
